+++ src/YesSql.Core/Services/TypeService.cs
@@ -1,6 +1,10 @@
 using System;
 using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
+using YesSql.Indexes;
+using YesSql.Serialization;
 
 namespace YesSql.Services
 {
@@ -9,6 +13,9 @@ namespace YesSql.Services
         private readonly ConcurrentDictionary<Type, string> typeNames = new();
 
         private readonly ConcurrentDictionary<string, Type> nameTypes = new();
+
+        private static readonly ConcurrentDictionary<PropertyInfo, PropertyInfoAccessor> PropertyAccessors = new();
+        private static readonly ConcurrentDictionary<Type, PropertyInfo[]> TypeProperties = new();
 
         public string this[Type t]
         {
@@ -55,6 +62,42 @@ namespace YesSql.Services
             return type.IsGenericType && type.Name.Contains("AnonymousType")
                    && (type.Name.StartsWith("<>") || type.Name.StartsWith("VB$"))
                    && (type.Attributes & TypeAttributes.NotPublic) == TypeAttributes.NotPublic;
+        }
+
+
+
+        public PropertyInfo[] GetProperties(Type type)
+        {
+            if (TypeProperties.TryGetValue(type, out var pis))
+            {
+                return pis;
+            }
+
+            var properties = type.GetProperties().Where(IsWriteable).ToArray();
+
+            var oldType = TypeProperties.FirstOrDefault(x => x.Key.FullName == type.FullName && x.Key != type);
+            if (oldType.Key != null)
+            {
+                TypeProperties.Remove(oldType.Key, out _);
+            }
+
+            TypeProperties[type] = properties;
+            return properties;
+        }
+
+        public PropertyInfoAccessor GetPropertyAccessors(PropertyInfo property, Func<PropertyInfo, PropertyInfoAccessor> createFactory)
+        {
+            return PropertyAccessors.GetOrAdd(property, createFactory(property));
+        }
+
+        private static bool IsWriteable(PropertyInfo pi)
+        {
+            return
+                pi.Name != nameof(IIndex.Id) &&
+                // don't read DocumentId when on a MapIndex as it might be used to 
+                // read the DocumentId directly from an Index query
+                pi.Name != "DocumentId"
+                ;
         }
     }
 }
