+++ test/YesSql.Tests/CoreTests.cs
@@ -50,7 +50,7 @@ namespace YesSql.Tests
 
                 _store = await StoreFactory.CreateAndInitializeAsync(_configuration);
                 await _store.InitializeCollectionAsync("Col1");
-                _store.TypeNames[typeof(Person)] = "People";
+                _store.TypeService[typeof(Person)] = "People";
 
                 await CreateTablesAsync(_configuration);
             }
@@ -58,7 +58,7 @@ namespace YesSql.Tests
             {
                 _store = await StoreFactory.CreateAndInitializeAsync(_configuration);
                 await _store.InitializeCollectionAsync("Col1");
-                _store.TypeNames[typeof(Person)] = "People";
+                _store.TypeService[typeof(Person)] = "People";
             }
 
             // Clear the tables for each new test
